Add files via upload
first prototype. still need to finalize what to do for the admin. also was thinking to set a boolean flag on the credentials table to identify who the admin is. It could also be a specific table dedicated to save the user IDs of the admins, but it doesn't inprove the security by much since once an hacker can freely access one table it is easy for him to move from table to table.